feat: aggregator json file generator
