+++ cypress_adapters.json
@@ -1,5 +1,5 @@
 {
-    "adapters": [
+    "result": [
         {
             "adapterHash": "0xf15e595e641f1447b5ad1c128007574b329e00937e6db3857aff1487345ed5f6",
             "name": "ADA-USDT",
